Fix capitalization of English pronouns
In German, capitaliazing "Du" and "Sie" is polite. But in English,
capitalizing "You" won't be seen as polite, just as incorrect grammar.

Note that legalese, such as found in licenses, is an exception. There,
the word "You" should remain capitalized since that usage has a specific
meaning in legal text. But in normal English text, "you" is correct.